refactor: rename auth module to user module
