+++ src/dependencies.py
@@ -2,6 +2,6 @@ from typing import Annotated
 
 from fastapi import Depends
 
-from src.auth.service import UserService
+from src.user.service import UserService
 
 UserServiceDependency = Annotated[UserService, Depends(UserService)]
